refac: reorganize wallet connection architecture
- Extract Providers to separate component
- Create dedicated ConnectButton component
- Add connectors module
- Remove legacy wallet hooks
