+++ src/components/Navbar.tsx
@@ -1,4 +1,9 @@
-import { ChevronDownIcon, EmailIcon } from '@chakra-ui/icons';
+'use client';
+
+import { useEffect } from 'react';
+import { useAccount } from '@starknet-react/core';
+import mixpanel from 'mixpanel-browser';
+import { usePathname } from 'next/navigation';
 import {
   Avatar,
   Box,
@@ -9,181 +14,23 @@ import {
   IconButton,
   Image,
   Link,
-  Menu,
-  MenuButton,
-  MenuItem,
-  MenuList,
   Text,
-  useDisclosure,
 } from '@chakra-ui/react';
 import { useSetAtom } from 'jotai';
-import {
-  ConnectOptionsWithConnectors,
-  StarknetkitConnector,
-  useStarknetkitConnectModal,
-  disconnect as starknetKitDisconnect,
-  connect,
-} from 'starknetkit';
 
-import argentMobile from '@/assets/argentMobile.svg';
 import tg from '@/assets/tg.svg';
 import CONSTANTS from '@/constants';
-import { useIsMobile } from '@/hooks/use-mobile';
 import { getERC20Balance } from '@/store/balance.atoms';
 import { addressAtom } from '@/store/claims.atoms';
-import {
-  getEndpoint,
-  getTokenInfoFromName,
-  MyMenuItemProps,
-  MyMenuListProps,
-  shortAddress,
-  standariseAddress,
-  truncate,
-} from '@/utils';
+import { getTokenInfoFromName, standariseAddress } from '@/utils';
 import fulllogo from '@public/fulllogo.svg';
-import {
-  InjectedConnector,
-  useAccount,
-  useConnect,
-  useDisconnect,
-  useStarkProfile,
-} from '@starknet-react/core';
-import mixpanel from 'mixpanel-browser';
-import { useEffect, useMemo } from 'react';
-import { constants } from 'starknet';
-import {
-  ArgentMobileConnector,
-  isInArgentMobileAppBrowser,
-} from 'starknetkit/argentMobile';
-import {
-  BraavosMobileConnector,
-  isInBraavosMobileAppBrowser,
-} from 'starknetkit/braavosMobile';
-import { WebWalletConnector } from 'starknetkit/webwallet';
 import TncModal from './TncModal';
+import { ConnectButton } from './ConnectButton';
 
-export function getConnectors(isMobile: boolean) {
-  const mobileConnector = ArgentMobileConnector.init({
-    options: {
-      dappName: 'Troves',
-      url: getEndpoint(),
-      chainId: constants.NetworkName.SN_MAIN,
-    },
-    inAppBrowserOptions: {},
-  }) as StarknetkitConnector;
-
-  const mobileBraavosConnector = BraavosMobileConnector.init({
-    inAppBrowserOptions: {},
-  }) as StarknetkitConnector;
-
-  const argentXConnector = new InjectedConnector({
-    options: {
-      id: 'argentX',
-      name: 'Argent X',
-    },
-  });
-
-  const braavosConnector = new InjectedConnector({
-    options: {
-      id: 'braavos',
-      name: 'Braavos',
-    },
-  });
-
-  const keplrConnector = new InjectedConnector({
-    options: {
-      id: 'keplr',
-      name: 'Keplr',
-    },
-  });
-
-  const isInstalled = [argentXConnector, braavosConnector, keplrConnector].map(
-    (wallet) => {
-      return {
-        id: wallet.id,
-        isInstalled:
-          typeof window === 'undefined'
-            ? false
-            : window[`starknet_${wallet.id}`] !== undefined,
-      };
-    },
-  );
-
-  const webWalletConnector = new WebWalletConnector({
-    url: 'https://web.argent.xyz',
-  }) as StarknetkitConnector;
-
-  if (isInArgentMobileAppBrowser()) {
-    return [mobileConnector];
-  } else if (isInBraavosMobileAppBrowser()) {
-    return [mobileBraavosConnector];
-  } else if (isMobile) {
-    return [
-      braavosConnector,
-      mobileConnector,
-      mobileBraavosConnector,
-      webWalletConnector,
-    ];
-  }
-
-  const defaultConnectors = [
-    argentXConnector,
-    braavosConnector,
-    keplrConnector,
-  ];
-
-  // put uninstall wallets at the end
-  const sortedConnectors: any[] = defaultConnectors.sort((a, b) => {
-    const aInstalled = isInstalled.find(
-      (wallet) => wallet.id === a.id,
-    )?.isInstalled;
-    const bInstalled = isInstalled.find(
-      (wallet) => wallet.id === b.id,
-    )?.isInstalled;
-
-    if (aInstalled && bInstalled) {
-      return 0;
-    } else if (aInstalled) {
-      return -1;
-    }
-    return 1;
-  });
-
-  sortedConnectors.push(mobileConnector);
-  sortedConnectors.push(webWalletConnector);
-  return sortedConnectors;
-}
-
-const walletIconMap: Record<string, any> = {
-  argentMobile,
-  argentWebWallet: EmailIcon,
-};
-
-const getWalletIcon = (walletId: string) => {
-  console.log(walletId, 'walletId');
-  return walletIconMap[walletId];
-};
-
-interface NavbarProps {
-  hideTg?: boolean;
-  forceShowConnect?: boolean;
-}
-
-export default function Navbar(props: NavbarProps) {
-  const { address, connector } = useAccount();
-  const { disconnectAsync } = useDisconnect();
+export default function Navbar() {
+  const pathname = usePathname();
+  const { address } = useAccount();
   const setAddress = useSetAtom(addressAtom);
-  const { data: starkProfile } = useStarkProfile({
-    address,
-    useDefaultPfp: true,
-  });
-  const { connect: connectSnReact } = useConnect();
-  const isMobile = useIsMobile();
-  const { starknetkitConnectModal } = useStarknetkitConnectModal({
-    connectors: getConnectors(isMobile) as StarknetkitConnector[],
-  });
-
-  // const [lastWallet, setLastWallet] = useAtom(lastWalletAtom);
 
   const getTokenBalance = async (token: string, address: string) => {
     const tokenInfo = getTokenInfoFromName(token);
@@ -191,50 +38,6 @@ export default function Navbar(props: NavbarProps) {
 
     return balance.amount.toEtherToFixedDecimals(6);
   };
-
-  const connectorConfig: ConnectOptionsWithConnectors = useMemo(() => {
-    return {
-      modalMode: 'alwaysAsk',
-      modalTheme: 'dark',
-      webWalletUrl: 'https://web.argent.xyz',
-      argentMobileOptions: {
-        dappName: 'Troves',
-        chainId: constants.NetworkName.SN_MAIN,
-        url: getEndpoint(),
-      },
-      dappName: 'Troves',
-      connectors: getConnectors(isMobile) as StarknetkitConnector[],
-    };
-  }, [isMobile]);
-
-  async function connectWallet(config = connectorConfig) {
-    try {
-      // const { connector } = await starknetkitConnectModal();
-      // if (!connector) {
-      //   return;
-      // }
-
-      // await connectSnReact({ connector: connector as any });
-
-      console.log(`connectWallet`, config);
-      const { connector } = await connect(config);
-      console.log(connector, 'connector');
-
-      if (connector) {
-        connectSnReact({ connector: connector as any });
-      }
-      return true;
-    } catch (error) {
-      console.error('connectWallet error', error);
-      return false;
-    }
-  }
-
-  useEffect(() => {
-    const config = connectorConfig;
-    console.log('connecting wallet');
-    connectWallet({ ...config, modalMode: 'neverAsk' });
-  }, []);
 
   useEffect(() => {
     (async () => {
@@ -253,22 +56,13 @@ export default function Navbar(props: NavbarProps) {
     })();
   }, [address]);
 
-  // Set last wallet when a new wallet is connected
-  // useEffect(() => {
-  //   console.log('lastWallet connector', connector?.name);
-  //   if (connector) {
-  //     const name: string = connector.name;
-  //     setLastWallet(name);
-  //   }
-  // }, [connector]);
-
   // set address atom
   useEffect(() => {
     console.log('tncinfo address', address);
     setAddress(address);
   }, [address]);
 
-  const { isOpen, onOpen, onClose } = useDisclosure();
+  const hideTg = pathname.includes('slinks');
 
   return (
     <Container
@@ -298,122 +92,33 @@ export default function Navbar(props: NavbarProps) {
               Troves
             </b>
           </span>
-          {/* <span style={{ display: 'flex', gap: '2px' }}>
-            There is an on-going network incident, transaction might fail.{' '}
-            <Link
-              href="https://x.com/Starknet/status/1962740091937317247"
-              textDecoration={'underline'}
-              target="_blank"
-            >
-              Learn more.
-            </Link>
-          </span> */}
         </Text>
       </Center>
       <Box
         width={'100%'}
         maxWidth="1152px"
         margin={'0px auto'}
-        padding={{ base: '20px 10px 10px' }}
+        padding={{ base: '20px 16px 10px', sm: '20px 10px 10px' }}
       >
-        <Flex width={'100%'} gap={2} justifyContent={'space-between'}>
-          <Link href="/" margin="auto 100px auto 0" textAlign={'left'}>
+        <Flex
+          width={'100%'}
+          gap={{ base: 1, sm: 2 }}
+          justifyContent={'space-between'}
+        >
+          <Link
+            href="/"
+            margin={{ base: 'auto 20px auto 0', sm: 'auto 100px auto 0' }}
+            textAlign={'left'}
+          >
             <Image
               src={fulllogo.src}
               alt="logo"
               height={{ base: '35px', md: '35px' }}
             />
-            {/* <Text fontSize={'30px'} color={'purple'} fontWeight={'bold'}>
-              Troves
-            </Text> */}
           </Link>
-          {/* <Link href={'/claims'} isExternal>
-            <Button
-              margin="0 0 0 auto"
-              borderColor="color2"
-              color="color2"
-              variant="ghost"
-              marginRight={'30px'}
-              leftIcon={
-                <Avatar
-                  size="sm"
-                  bg="highlight"
-                  color="color2"
-                  name="T G"
-                  src={CONSTANTS.LOGOS.STRK}
-                />
-              }
-              _hover={{
-                bg: 'color2_50p',
-              }}
-              display={{ base: 'none !important', md: 'flex !important' }}
-            >
-              Claims
-            </Button>
-          </Link> */}
-
-          {/* <Link
-            href="/"
-            display={'flex'}
-            alignItems={'center'}
-            marginRight="auto"
-          >
-            <Button
-              bg="transparent"
-              color="text_secondary"
-              variant="outline"
-              border="none"
-              px="10px"
-              _hover={{
-                color: 'text_primary',
-              }}
-              display={{ base: 'none !important', lg: 'flex !important' }}
-              onClick={() => {
-                mixpanel.track('home_clicked');
-              }}
-            >
-              Home
-            </Button>
-          </Link> */}
-          {/* <Link href="/raffle" margin="0 10px 0 0">
-            <Button
-              bg="transparent"
-              color="color2"
-              variant="outline"
-              border="none"
-              px="10px"
-              _hover={{
-                bg: 'color2_50p',
-              }}
-              display={{ base: 'none !important', lg: 'flex !important' }}
-              onClick={() => {
-                mixpanel.track('home_clicked');
-              }}
-            >
-              🕹 {'  '}Raffle
-            </Button>
-          </Link> */}
-          {/* <Link href="/community" display={'flex'} alignItems={'center'}>
-            <Button
-              bg="transparent"
-              color="text_secondary"
-              variant="outline"
-              border="none"
-              _hover={{
-                color: 'text_primary',
-              }}
-              px="10px"
-              display={{ base: 'none !important', lg: 'flex !important' }}
-              onClick={() => {
-                mixpanel.track('community_program_click');
-              }}
-            >
-              ✨ Community Program
-            </Button>
-          </Link> */}
 
           <Flex gap={2}>
-            {!props.hideTg && (
+            {!hideTg && (
               <Link
                 href={CONSTANTS.COMMUNITY_TG}
                 textDecoration="none !important"
@@ -466,126 +171,7 @@ export default function Navbar(props: NavbarProps) {
               </Link>
             )}
 
-            {true && (
-              <Box display={'flex'} alignItems={'center'}>
-                <Menu>
-                  <MenuButton
-                    as={Button}
-                    rightIcon={address ? <ChevronDownIcon /> : <></>}
-                    iconSpacing={{ base: '1px', sm: '5px' }}
-                    background="connect_button_gradient"
-                    color={'black'}
-                    borderRadius={'8px'}
-                    display={{ base: 'flex' }}
-                    height={{ base: '2rem', sm: '2.5rem' }}
-                    my={{ base: 'auto', sm: 'initial' }}
-                    paddingX={{ base: '0.5rem', sm: '1rem' }}
-                    fontSize={{ base: '0.8rem', sm: '0.8rem' }}
-                    fontWeight={'bold'}
-                    _hover={{
-                      background: 'purple_hover_2',
-                    }}
-                    _active={{
-                      bg: 'purple_hover_2',
-                    }}
-                    onClick={
-                      address
-                        ? undefined
-                        : () => {
-                            connectWallet();
-                          }
-                    }
-                  >
-                    <Center>
-                      {address ? (
-                        <Center display="flex" alignItems="center" gap=".5rem">
-                          <Image
-                            src={
-                              starkProfile?.profilePicture ||
-                              connector?.id === 'argentMobile'
-                                ? getWalletIcon(connector?.id ?? '')?.src ||
-                                  connector?.icon.toString() ||
-                                  '/fallback-profile-icon.jpeg'
-                                : '/fallback-profile-icon.jpeg'
-                            }
-                            alt="pfp"
-                            width={{ base: '20px', sm: '22px' }}
-                            height={{ base: '20px', sm: '22px' }}
-                            rounded="full"
-                            background={'mybg'}
-                            padding={'3px'}
-                          />{' '}
-                          <Text as="h3" marginTop={'3px !important'}>
-                            {starkProfile && starkProfile.name
-                              ? truncate(starkProfile.name, 6, isMobile ? 0 : 6)
-                              : shortAddress(address, 4, isMobile ? 0 : 4)}
-                          </Text>
-                        </Center>
-                      ) : (
-                        'Connect wallet'
-                      )}
-                    </Center>
-                  </MenuButton>
-                  <MenuList {...MyMenuListProps}>
-                    {address && (
-                      <MenuItem
-                        {...MyMenuItemProps}
-                        onClick={() => {
-                          disconnectAsync().then((data) => {
-                            console.log('wallet disconnected');
-                            // setLastWallet(null);
-                            starknetKitDisconnect({ clearLastWallet: true });
-                          });
-                        }}
-                      >
-                        Disconnect
-                      </MenuItem>
-                    )}
-                  </MenuList>
-                </Menu>
-              </Box>
-            )}
-            {/* 
-            {isMobile && (
-              <IconButton
-                aria-label="Open menu"
-                icon={<HamburgerIcon color="purple" height="30px" width="30px" />}
-                background="transparent"
-                display={{ base: 'flex', md: 'none' }}
-                onClick={onOpen}
-                _focus={{
-                  bg: 'none',
-                }}
-              />
-            )} */}
-
-            {/* <Drawer placement="right" onClose={onClose} isOpen={isOpen}>
-              <DrawerOverlay />
-              <DrawerContent background="mycard">
-                <DrawerHeader color="text_secondary">Menu</DrawerHeader>
-                <DrawerBody>
-                  <Flex direction="column">
-                    <Link href="/" color="text_secondary" onClick={onClose}>
-                      Home
-                    </Link>
-                    <Link href="/raffle" color="text_secondary" onClick={onClose}>
-                      🕹 {'  '}Raffle
-                    </Link>
-                    <Link
-                      href="/community"
-                      color="text_secondary"
-                      onClick={() => {
-                        onClose();
-                        mixpanel.track('community_program_click');
-                      }}
-                      mt={4}
-                    >
-                      ✨ Community Program
-                    </Link>
-                  </Flex>
-                </DrawerBody>
-              </DrawerContent>
-            </Drawer> */}
+            <ConnectButton />
           </Flex>
         </Flex>
       </Box>
